Separate framework projects
- Gaois.QueryLogger now multi-targets .NET Framework and .NET Standard
- This allows us to use the native .NET Framework HTTPContext to access host names, client IP addresses, etc.
- At the same time, the .NET Standard implementation can be referenced by a new library that targets .NET Core specifically, allowing us to implement equivalent functionality there.
- Until the .NET Core library is completed, .NET Core support is broken in this build.
