+++ src/Gaois.QueryLogger.AspNetCore/Gaois.QueryLogger.AspNetCore.csproj
@@ -1,0 +1,14 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore" Version="2.1.0" />
+    <Reference Include="Gaois.QueryLogger">
+      <HintPath>..\Gaois.QueryLogger\bin\Debug\net461\Gaois.QueryLogger.dll</HintPath>
+    </Reference>
+  </ItemGroup>
+
+</Project>
