+++ src/Gaois.QueryLogger/Gaois.QueryLogger.nuspec
