+++ src/Gaois.QueryLogger/IPAddressAnonymizationLevel.cs
