+++ src/Gaois.QueryLogger/LogStore.Async.cs
