+++ src/Gaois.QueryLogger/LogStore.cs
