+++ src/Gaois.QueryLogger/Query.cs
@@ -55,7 +55,7 @@ namespace Gaois.QueryLogger
         /// <summary>
         /// The date and time of query logging. Gaois.QueryLogger logs this information automatically.
         /// </summary>
-        public DateTime LogDate { get; set; }
+        public DateTime? LogDate { get; set; }
 
         /// <summary>
         /// Record additional data in JSON format
