+++ src/Gaois.QueryLogger/QueryLogger.Async.cs
