+++ src/Gaois.QueryLogger/QueryLoggerSettings.cs
